+++ src/main/java/com/revature/RpmAuthServiceApplication.java
@@ -1,4 +1,4 @@
-package com.revature.rpmauthservice;
+package com.revature;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
