--- src/main/java/com/revature/rpmauthservice/controllers/AuthController.java
@@ -1,13 +1,0 @@
-package com.revature.rpmauthservice.controllers;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class AuthController {
-
-	@RequestMapping("/")
-	public String home() {
-		return "best string";
-	}
-}
